Added the basic recipe
This adds the basic recipe, which includes the examples and source
code.

The data for the state select options have been placed into a
separate file so they can be included anywhere.

The main page has also been updated so the "View" link is placed
within the header of the panel.